refactor: Table, Order간의 양방향 의존 관계 제거
