+++ src/main/java/kitchenpos/application/table/TableGroupingService.java
@@ -1,0 +1,8 @@
+package kitchenpos.application.table;
+
+public interface TableGroupingService {
+
+    void isAbleToGroup(final Long orderTableId);
+
+    void isAbleToUngroup(final Long orderTableId);
+}
