+++ src/main/java/kitchenpos/application/table/TableService.java
@@ -6,9 +6,7 @@ import kitchenpos.application.dto.OrderTableCreationRequest;
 import kitchenpos.application.dto.OrderTableEmptyStatusChangeRequest;
 import kitchenpos.application.dto.OrderTableGuestAmountChangeRequest;
 import kitchenpos.application.dto.result.OrderTableResult;
-import kitchenpos.dao.order.OrderRepository;
 import kitchenpos.dao.table.OrderTableRepository;
-import kitchenpos.domain.order.Order;
 import kitchenpos.domain.table.OrderTable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,11 +15,14 @@ import org.springframework.transaction.annotation.Transactional;
 public class TableService {
 
     private final OrderTableRepository orderTableRepository;
-    private final OrderRepository orderRepository;
+    private final TableGroupingService tableGroupingService;
 
-    public TableService(final OrderTableRepository orderTableRepository, final OrderRepository orderRepository) {
+    public TableService(
+            final OrderTableRepository orderTableRepository,
+            final TableGroupingService tableGroupingService
+    ) {
         this.orderTableRepository = orderTableRepository;
-        this.orderRepository = orderRepository;
+        this.tableGroupingService = tableGroupingService;
     }
 
     @Transactional
@@ -41,17 +42,9 @@ public class TableService {
     public OrderTableResult changeEmpty(final Long orderTableId, final OrderTableEmptyStatusChangeRequest request) {
         final OrderTable orderTable = orderTableRepository.findById(orderTableId)
                 .orElseThrow(() -> new IllegalArgumentException("Order table does not exist."));
-        final List<Order> orders = orderRepository.findAllByOrderTableId(orderTableId);
-        validateOrdersStatus(orders);
+        tableGroupingService.isAbleToGroup(orderTableId);
         orderTable.changeEmpty(request.getEmpty());
         return OrderTableResult.from(orderTableRepository.save(orderTable));
-    }
-
-    private void validateOrdersStatus(final List<Order> orders) {
-        if (orders.stream().allMatch(Order::isCompleted)) {
-            return;
-        }
-        throw new IllegalArgumentException("Cannot change empty status of table with order status not completion");
     }
 
     @Transactional
